api (deletebrand & deletecategory) - delete only unassociated brands

+++ pages/api/brands/deletebrand.js
@@ -1,4 +1,5 @@
 import Brand from "@/models/Brand";
+import Product from "@/models/Product";
 import connectDb from "@/db/mongoose";
 
 const handler = async (req, res) => {
@@ -6,7 +7,19 @@ const handler = async (req, res) => {
     const { _id } = req.query;
 
     try {
-      // Find the brand by id and delete it
+      // Check if there are any associated products
+      const productsCount = await Product.countDocuments({ brand: _id });
+
+      if (productsCount > 0) {
+        return res
+          .status(400)
+          .json({
+            success: false,
+            error: "Cannot delete brand. Associated products exist",
+          });
+      }
+
+      // Delete the brand
       const brand = await Brand.findByIdAndDelete(_id);
 
       if (!brand) {
@@ -14,10 +27,12 @@ const handler = async (req, res) => {
           .status(404)
           .json({ success: false, error: "Brand not found" });
       }
+
       res
         .status(200)
         .json({ success: true, message: "Brand deleted successfully" });
     } catch (error) {
+      console.error(error); // Log the error for debugging purposes
       res.status(500).json({ success: false, error: "Internal server error" });
     }
   } else {
